Fjerner gamle komponenter
