+++ src/index.less
@@ -35,6 +35,11 @@ body.no-scroll {
     background-color: white;
 }
 
+.desktopmeny {
+    height: 106px;
+    padding: 1rem 0;
+}
+
 .header-z-wrapper {
     z-index: 1000;
     position: relative;
