+++ src/komponenter/common/nav-logo/NavLogoLenke.less
@@ -1,0 +1,25 @@
+@import '../../../styling-variabler.less';
+
+.nav-logo-lenke {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    svg {
+        width: 100%;
+        height: 100%;
+    }
+
+    &:hover {
+        svg {
+            transform: scale(1.05);
+        }
+    }
+
+    &:focus {
+        box-shadow: 0 0 0 4px @fokusFarge inset;
+        border-radius: 8px;
+        outline: none;
+        border-color: transparent;
+    }
+}
