+++ src/komponenter/footer/footer-regular/footer-bottom/FooterBottom.less
@@ -16,12 +16,12 @@
     width: 100%;
     max-width: 1344px;
 
-    &__bottom-logo {
-        height: 64px;
-        width: 64px;
-        &:hover {
-            height: 68px;
-            width: 68px;
+    .nav-logo-lenke {
+        width: 68px;
+        padding: 0.25rem;
+        margin-bottom: 0.25rem;
+        .nav-logo-fill {
+            fill: black;
         }
     }
