--- src/komponenter/footer/footer-regular/footer-bottom/logo/NavLogoFooter.less
@@ -1,37 +1,0 @@
-@import (reference) '../../../../../styling-variabler.less';
-
-.sitefooter {
-    .navbar-brand {
-        width: 68px;
-        height: 68px;
-
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-right: 0.8rem;
-        border-radius: 4px;
-
-        img {
-            width: 100%;
-        }
-
-        &:hover {
-            img {
-                transform: scale(1.05);
-            }
-        }
-
-        &:focus,
-        &:active {
-            box-shadow: 0 0 0 4px @fokusFarge;
-            background-color: transparent;
-            outline: none;
-            border-color: transparent;
-        }
-
-        @media (min-width: @desktopMinWidth) {
-            width: 65px;
-            height: 65px;
-        }
-    }
-}
