+++ src/komponenter/header/header-regular/desktop/DesktopMenylinje.test.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import DesktopMenylinje from './DesktopMenylinje';
-import NavLogoHeader from 'komponenter/header/header-regular/common/logo/NavLogoHeader';
 import HovedmenyDesktop from './hovedmeny/HovedmenyDesktop';
 import { SokDropdown } from './sok/SokDropdown';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import { reducers } from 'store/reducers';
 import { mount } from 'enzyme';
+import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 
 const store = createStore(reducers);
 
@@ -18,9 +18,9 @@ const shallowWithProps = () =>
     );
 
 describe('<DesktopMenylinje>', () => {
-    it('Skal rendre <NavLogoHeader> komponent', () => {
+    it('Skal rendre <NavLogoLenke> komponent', () => {
         const wrapper = shallowWithProps();
-        expect(wrapper.find(NavLogoHeader)).toHaveLength(1);
+        expect(wrapper.find(NavLogoLenke)).toHaveLength(1);
     });
 
     it('Skal rendre <HovedmenyDesktop> komponent hvis språk er norsk', () => {
