+++ src/komponenter/header/header-regular/mobil/MobilMenylinje.less
@@ -37,6 +37,11 @@
         }
     }
 
+    .nav-logo-lenke {
+        width: 72px;
+        padding: 4px;
+    }
+
     .login-knapp-container {
         order: 2;
         margin: 0 3px;
