+++ src/komponenter/header/header-regular/mobil/MobilMenylinje.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import BEMHelper from 'utils/bem';
-import NavLogoHeader from 'komponenter/header/header-regular/common/logo/NavLogoHeader';
 import { Language } from 'store/reducers/language-duck';
 import { useSelector } from 'react-redux';
 import { AppState } from 'store/reducers';
@@ -9,6 +8,8 @@ import { MenuValue } from 'utils/meny-storage-utils';
 import { Status } from 'api/api';
 import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
 import { Varselbjelle } from 'komponenter/header/header-regular/mobil/varsler/Varselbjelle';
+import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
+import { GACategory } from 'utils/google-analytics';
 import './MobilMenylinje.less';
 
 const mobilClass = BEMHelper('mobilmeny');
@@ -48,7 +49,13 @@ const MobilMenylinje = ({ language }: Props) => {
             aria-label="Hovedmeny"
         >
             <div className={mobilClass.element('elementer')}>
-                <NavLogoHeader />
+                <NavLogoLenke
+                    gaEventArgs={{
+                        context: arbeidsflate,
+                        category: GACategory.Header,
+                        action: 'navlogo',
+                    }}
+                />
                 <div className={mobilClass.element('hoyre-kolonne')}>
                     {visInnloggingsKnapp && <LoggInnKnapp type={'flat'} />}
                     {visVarslerDropdown && <Varselbjelle />}
