+++ src/komponenter/header/header-simple/HeaderSimple.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import BEMHelper from 'utils/bem';
-import NavLogoSimple from 'komponenter/header/header-simple/logo/NavLogoSimple';
+import NavLogoLenke from 'komponenter/common/nav-logo/NavLogoLenke';
 import Navn from './navn/Navn';
 import LoggInnKnapp from 'komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp';
 import { Sticky } from 'komponenter/header/header-regular/common/sticky/Sticky';
+import { GACategory } from 'utils/google-analytics';
 import './HeaderSimple.less';
 
 export const SimpleHeader = () => {
@@ -13,7 +14,12 @@ export const SimpleHeader = () => {
         <Sticky>
             <div className={cls.className}>
                 <div className={cls.element('content')}>
-                    <NavLogoSimple className={cls.element('logo')} />
+                    <NavLogoLenke
+                        gaEventArgs={{
+                            category: GACategory.Header,
+                            action: 'navlogo-mobilmeny',
+                        }}
+                    />
                     <div className={cls.element('right')}>
                         <Navn />
                         <LoggInnKnapp />
